requirements.txt
pytesseract
pynput
Pillow
